// src/extension/types/startgg-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  JSON: { input: any; output: any; }
  Timestamp: { input: any; output: any; }
};
// ...
export type Event = {
  __typename?: 'Event';
  /** How long before the event start will the check-in end (in seconds) */
  checkInBuffer?: Maybe<Scalars['Int']['output']>;
  /** How long the event check-in will last (in seconds) */
  checkInDuration?: Maybe<Scalars['Int']['output']>;
  /** Whether check-in is enabled for this event */
  checkInEnabled?: Maybe<Scalars['Boolean']['output']>;
  /** Rough categorization of event tier, denoting relative importance in the competitive scene */
  competitionTier?: Maybe<Scalars['Int']['output']>;
  /** When the event was created (unix timestamp) */
  createdAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Last date attendees are able to create teams for team events */
  deckSubmissionDeadline?: Maybe<Scalars['Timestamp']['output']>;
  /**
   * Maximum number of participants each Entrant can have
   * @deprecated Migrate to teamRosterSize
   */
  entrantSizeMax?: Maybe<Scalars['Int']['output']>;
  /**
   * Minimum number of participants each Entrant can have
   * @deprecated Migrate to teamRosterSize
   */
  entrantSizeMin?: Maybe<Scalars['Int']['output']>;
  /** The entrants that belong to an event, paginated by filter criteria */
  entrants?: Maybe<EntrantConnection>;
  /** Whether the event has decks */
  hasDecks?: Maybe<Scalars['Boolean']['output']>;
  /** Are player tasks enabled for this event */
  hasTasks?: Maybe<Scalars['Boolean']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  /** Whether the event is an online event or not */
  isOnline?: Maybe<Scalars['Boolean']['output']>;
  league?: Maybe<League>;
  /** Markdown field for match rules/instructions */
  matchRulesMarkdown?: Maybe<Scalars['String']['output']>;
  /** Title of event set by organizer */
  name?: Maybe<Scalars['String']['output']>;
  /** Gets the number of entrants in this event */
  numEntrants?: Maybe<Scalars['Int']['output']>;
  /** The phase groups that belong to an event. */
  phaseGroups?: Maybe<Array<Maybe<PhaseGroup>>>;
  /** The phases that belong to an event. */
  phases?: Maybe<Array<Maybe<Phase>>>;
  /** TO settings for prizing */
  prizingInfo?: Maybe<Scalars['JSON']['output']>;
  publishing?: Maybe<Scalars['JSON']['output']>;
  /** Markdown field for event rules/instructions */
  rulesMarkdown?: Maybe<Scalars['String']['output']>;
  /** Id of the event ruleset */
  rulesetId?: Maybe<Scalars['Int']['output']>;
  /**
   * Settings pulled from the event ruleset, if one exists
   * @deprecated Use ruleset
   */
  rulesetSettings?: Maybe<Scalars['JSON']['output']>;
  /** Paginated sets for this Event */
  sets?: Maybe<SetConnection>;
  slug?: Maybe<Scalars['String']['output']>;
  /** Paginated list of standings */
  standings?: Maybe<StandingConnection>;
  /** When does this event start? */
  startAt?: Maybe<Scalars['Timestamp']['output']>;
  /** The state of the Event. */
  state?: Maybe<ActivityState>;
  /** Paginated stations on this event */
  stations?: Maybe<StationsConnection>;
  /** Last date attendees are able to create teams for team events */
  teamManagementDeadline?: Maybe<Scalars['Timestamp']['output']>;
  /** If this is a teams event, returns whether or not teams can set custom names */
  teamNameAllowed?: Maybe<Scalars['Boolean']['output']>;
  /** Team roster size requirements */
  teamRosterSize?: Maybe<TeamRosterSize>;
  tournament?: Maybe<Tournament>;
  /** The type of the event, whether an entrant will have one participant or multiple */
  type?: Maybe<Scalars['Int']['output']>;
  /** When the event was last modified (unix timestamp) */
  updatedAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Whether the event uses the new EventSeeds for seeding */
  useEventSeeds?: Maybe<Scalars['Boolean']['output']>;
  /** The entrant (if applicable) for a given user in this event */
  userEntrant?: Maybe<Entrant>;
  videogame?: Maybe<Videogame>;
  /** The waves being used by the event */
  waves?: Maybe<Array<Maybe<Wave>>>;
};
// ...
export type EventConnection = {
  __typename?: 'EventConnection';
  nodes?: Maybe<Array<Maybe<Event>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type Image = {
  __typename?: 'Image';
  height?: Maybe<Scalars['Float']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  ratio?: Maybe<Scalars['Float']['output']>;
  type?: Maybe<Scalars['String']['output']>;
  url?: Maybe<Scalars['String']['output']>;
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type LeagueConnection = {
  __typename?: 'LeagueConnection';
  nodes?: Maybe<Array<Maybe<League>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type ParticipantConnection = {
  __typename?: 'ParticipantConnection';
  nodes?: Maybe<Array<Maybe<Participant>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type ProfileAuthorization = {
  __typename?: 'ProfileAuthorization';
  /** The id given by the external service */
  externalId?: Maybe<Scalars['String']['output']>;
  /** The username given by the external service (including discriminator if discord) */
  externalUsername?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  stream?: Maybe<Stream>;
  /** The name of the external service providing this auth i.e. "twitch" */
  type?: Maybe<AuthorizationType>;
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type StationsConnection = {
  __typename?: 'StationsConnection';
  nodes?: Maybe<Array<Maybe<Stations>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type StreamQueue = {
  __typename?: 'StreamQueue';
  id?: Maybe<Scalars['String']['output']>;
  /** The sets on the stream */
  sets?: Maybe<Array<Maybe<Set>>>;
  /** The stream on the queue */
  stream?: Maybe<Streams>;
};
// ...
export type Streams = {
  __typename?: 'Streams';
  enabled?: Maybe<Scalars['Boolean']['output']>;
  followerCount?: Maybe<Scalars['Int']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  isOnline?: Maybe<Scalars['Boolean']['output']>;
  numSetups?: Maybe<Scalars['Int']['output']>;
  parentStreamId?: Maybe<Scalars['Int']['output']>;
  streamGame?: Maybe<Scalars['String']['output']>;
  streamId?: Maybe<Scalars['String']['output']>;
  streamLogo?: Maybe<Scalars['String']['output']>;
  streamName?: Maybe<Scalars['String']['output']>;
  streamSource?: Maybe<StreamSource>;
  streamStatus?: Maybe<Scalars['String']['output']>;
  streamType?: Maybe<Scalars['Int']['output']>;
  streamTypeId?: Maybe<Scalars['Int']['output']>;
};
// ...
export type TeamConnection = {
  __typename?: 'TeamConnection';
  nodes?: Maybe<Array<Maybe<Team>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type Tournament = {
  __typename?: 'Tournament';
  addrState?: Maybe<Scalars['String']['output']>;
  /** Admin-only view of admins for this tournament */
  admins?: Maybe<Array<Maybe<User>>>;
  city?: Maybe<Scalars['String']['output']>;
  countryCode?: Maybe<Scalars['String']['output']>;
  /** When the tournament was created (unix timestamp) */
  createdAt?: Maybe<Scalars['Timestamp']['output']>;
  currency?: Maybe<Scalars['String']['output']>;
  /** When the tournament ends */
  endAt?: Maybe<Scalars['Timestamp']['output']>;
  /** When does event registration close */
  eventRegistrationClosesAt?: Maybe<Scalars['Timestamp']['output']>;
  events?: Maybe<Array<Maybe<Event>>>;
  /** True if tournament has at least one offline event */
  hasOfflineEvents?: Maybe<Scalars['Boolean']['output']>;
  hasOnlineEvents?: Maybe<Scalars['Boolean']['output']>;
  hashtag?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  /** True if tournament has at least one online event */
  isOnline?: Maybe<Scalars['Boolean']['output']>;
  /** Is tournament registration open */
  isRegistrationOpen?: Maybe<Scalars['Boolean']['output']>;
  lat?: Maybe<Scalars['Float']['output']>;
  links?: Maybe<TournamentLinks>;
  lng?: Maybe<Scalars['Float']['output']>;
  mapsPlaceId?: Maybe<Scalars['String']['output']>;
  /** The tournament name */
  name?: Maybe<Scalars['String']['output']>;
  /** Number of attendees including spectators, if public */
  numAttendees?: Maybe<Scalars['Int']['output']>;
  /** The user who created the tournament */
  owner?: Maybe<User>;
  /** Paginated, queryable list of participants */
  participants?: Maybe<ParticipantConnection>;
  postalCode?: Maybe<Scalars['String']['output']>;
  primaryContact?: Maybe<Scalars['String']['output']>;
  primaryContactType?: Maybe<Scalars['String']['output']>;
  /** Publishing settings for this tournament */
  publishing?: Maybe<Scalars['JSON']['output']>;
  /** When does registration for the tournament end */
  registrationClosesAt?: Maybe<Scalars['Timestamp']['output']>;
  rules?: Maybe<Scalars['String']['output']>;
  /** The short slug used to form the url */
  shortSlug?: Maybe<Scalars['String']['output']>;
  /** The slug used to form the url */
  slug?: Maybe<Scalars['String']['output']>;
  /** When the tournament Starts */
  startAt?: Maybe<Scalars['Timestamp']['output']>;
  /** State of the tournament, can be ActivityState::CREATED, ActivityState::ACTIVE, or ActivityState::COMPLETED */
  state?: Maybe<Scalars['Int']['output']>;
  stations?: Maybe<StationsConnection>;
  streamQueue?: Maybe<Array<Maybe<StreamQueue>>>;
  streams?: Maybe<Array<Maybe<Streams>>>;
  /** When is the team creation deadline */
  teamCreationClosesAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Paginated, queryable list of teams */
  teams?: Maybe<TeamConnection>;
  /** The timezone of the tournament */
  timezone?: Maybe<Scalars['String']['output']>;
  /** The type of tournament from TournamentType */
  tournamentType?: Maybe<Scalars['Int']['output']>;
  /** When the tournament was last modified (unix timestamp) */
  updatedAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Build Tournament URL */
  url?: Maybe<Scalars['String']['output']>;
  venueAddress?: Maybe<Scalars['String']['output']>;
  venueName?: Maybe<Scalars['String']['output']>;
  /** List of all waves in this tournament */
  waves?: Maybe<Array<Maybe<Wave>>>;
};
// ...
export type TournamentLinks = {
  __typename?: 'TournamentLinks';
  discord?: Maybe<Scalars['String']['output']>;
  facebook?: Maybe<Scalars['String']['output']>;
};
// ...
export type User = {
  __typename?: 'User';
  /** Authorizations to external services (i.e. Twitch, Twitter) */
  authorizations?: Maybe<Array<Maybe<ProfileAuthorization>>>;
  bio?: Maybe<Scalars['String']['output']>;
  /** Public facing user birthday that respects user publishing settings */
  birthday?: Maybe<Scalars['String']['output']>;
  /** Uniquely identifying token for user. Same as the hashed part of the slug */
  discriminator?: Maybe<Scalars['String']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  /** Events this user has competed in */
  events?: Maybe<EventConnection>;
  genderPronoun?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  /** Leagues this user has competed in */
  leagues?: Maybe<LeagueConnection>;
  /** Public location info for this user */
  location?: Maybe<Address>;
  /** Public facing user name that respects user publishing settings */
  name?: Maybe<Scalars['String']['output']>;
  /** player for user */
  player?: Maybe<Player>;
  slug?: Maybe<Scalars['String']['output']>;
  /** Tournaments this user is organizing or competing in */
  tournaments?: Maybe<TournamentConnection>;
};
// ...
export type Wave = {
  __typename?: 'Wave';
  id?: Maybe<Scalars['ID']['output']>;
  /** The Wave Identifier */
  identifier?: Maybe<Scalars['String']['output']>;
  /** Unix time the wave is scheduled to start. */
  startAt?: Maybe<Scalars['Timestamp']['output']>;
};